fix: harden input validation, warnings, and error reporting
- Replace atof/atoi option parsing with strict strtod/strtol helpers so
  malformed numeric arguments (e.g. `-f abc`) are rejected instead of being
  silently coerced to 0
- Add the missing `index + 1 < argc` guard to `-b`, fixing an out-of-bounds
  argv read when `-b` is the final argument
- Return a non-zero exit status on error (previously main() always returned 0),
  so scripts and tests can detect failures; `--help`/`--version` still exit 0
- Preserve the first error instead of letting finish_wave_file() overwrite it
  when writing a .wav file, which was masking failures
- Fix -Wextra warnings: signed/unsigned comparisons in patterns.c, and use the
  previously-unused file_size in play_wav_data() as a defensive bound on
  untrusted WAV input

+++ patterns.c
@@ -226,7 +226,7 @@ SoundError play_code(double freq, double dit, bool paris_standard, double farnsw
         gap_dit = dit + extra / GAP_DITS;
     }
 
-    for (int k = 0; k < strlen(text); k++) {
+    for (size_t k = 0; k < strlen(text); k++) {
         unsigned char c = (unsigned char)toupper(text[k]);
         char sequence[16];
 
@@ -357,7 +357,7 @@ SoundError play_code(double freq, double dit, bool paris_standard, double farnsw
             }
         }
 
-        for (int i = 0; i < strlen(sequence); i++) {
+        for (size_t i = 0; i < strlen(sequence); i++) {
             if (sequence[i] == '~') {
                 fill_buffer_or_file(SILENCE, 1000.0, out_file);
